requirements.txt
Flask==0.12.4
Flask-LogConfig==0.4.2
Flask-Script==2.0.5
requests==2.20.0
pycryptodome==3.7.0
SQLAlchemy==1.0.13
Flask-SQLAlchemy==2.1
Flask-Migrate==1.8.0
alembic==0.8.6
psycopg2==2.6.2
git+https://github.com/LandRegistry/jwt-validation.git@v1.0.0
